Revert "Main loop update."
This reverts commit cf0fd42dab6c86ef9cba0200575ddb5fb1f41234.

+++ TestConsoleApp/Neurone.cs
@@ -9,7 +9,7 @@ namespace TestConsoleApp
         {
         }
 
-        public float NeuroneFunction(List<float> input, List<float> weight, float bias, string function)
+        public double NeuroneFunction(List<float> input, List<float> weight, float bias, string function)
         {
             float result = 0;
             for (int i = 0; i <= input.Count; i++)
